Separate out api token tests
